FIX: rearranging types and fixing last type issues preventing compile.
