+++ src/api/sp1/plonk.ts
@@ -1,7 +1,7 @@
 import { Sp1PlonkComputationalPlan } from '../../compute/plans/sp1/plonk.js';
 import { Logger } from 'esm-iso-logger';
 import { ApiMethod } from '../methodDecorator.js';
-import { sp1PlonkObjKeys, type Sp1Input, type Sp1PlonkInputProcessed } from './types.js';
+import { sp1PlonkObjKeys, type Sp1Input, type Sp1PlonkInputTransformed } from "../validation/sp1/schema.js";
 import { assertExactStructure } from '../validation/validation.js';
 import { sp1PlonkInputSchema } from '../validation/sp1/schema.js';
 
@@ -19,7 +19,7 @@ const fromSp1Object = (obj: unknown) => {
 };
 
 export const performSp1Plonk = ApiMethod<
-  Sp1PlonkInputProcessed, // TInput: processed shape given to executor
+  Sp1PlonkInputTransformed, // TInput: processed shape given to executor
   // Disabling args at this time as unknown impact on new TEE option the decoding might change!
   false, //typeof sp1ArgKeys, // TKeys (what arguments mode expects to be provided) performSp1ToPlonk.fromArgs(hexPi, programVK, encodedProof)
   Sp1Input // TObject (what object mode expects as a single object) performSp1Plonk.fromObject({} as Sp1Input)
