+++ src/api/validation/risc0/schema.ts
@@ -1,6 +1,21 @@
+import type { O1jsVK, O1jsProof } from "pairing-utils/pkg/pairing_utils.js";
 import { isAffinePoint2d, isComplexAffinePoint2d, isField12, isString } from "../guards.js";
 
-// Helper validators for complex nested structures matching pairing_utils types
+// Types ======================================================================================
+
+export type Risc0Groth16Input = {
+  risc0_proof: Risc0Groth16Proof;
+  raw_vk: Risc0Groth16Vk;
+};
+
+// Original vk risc0 gives us - Must include all O1jsVK keys but must exclude ic6 AND alpha_beta - aka we must have 5 PIs starting from index 1 (pi1->pi5) - pi0 is the groth16 verification key itself
+export type Risc0Groth16Vk = Omit<Required<O1jsVK>, 'ic6' | 'alpha_beta'>;
+// Must include all O1jsVK keys but must exclude ic6 - aka we must have 6 ICs starting from index 0 (ic0->ic5)
+export type Risc0Groth16PairedVk = Omit<Required<O1jsVK>, 'ic6'>;
+// Must include all O1jsProof keys but must exclude pi6 - aka we must have 5 PIs starting from index 1 (pi1->pi5) - pi0 is the groth16 verification key itself
+export type Risc0Groth16Proof = Omit<Required<O1jsProof>, 'pi6'>;
+
+// Runtime validation =========================================================================
 
 const risc0Groth16ProofSchema = {
   negA: isAffinePoint2d,
@@ -27,11 +42,12 @@ const risc0Groth16RawVkSchema = {
   ic5: isAffinePoint2d,
 };
 
-// Schema for args form (just the proof) - what crap is this NO IT NOT FIXME
-export const risc0Groth16ArgsInputSchema = risc0Groth16ProofSchema;
-
 // Schema for obj form (risc0_proof + raw_vk)
 export const risc0Groth16ObjInputSchema = {
   risc0_proof: risc0Groth16ProofSchema,
   raw_vk: risc0Groth16RawVkSchema,
 };
+
+// Keys for the ApiMethod helper
+export const risc0Groth16ArgsKeys = Object.keys(risc0Groth16ObjInputSchema) as (keyof Risc0Groth16Input)[];
+export const risc0Groth16ObjKeys = Object.keys(risc0Groth16ObjInputSchema) as (keyof Risc0Groth16Input)[];
