+++ src/index.ts
@@ -7,7 +7,7 @@ export { InvertedPromise } from './utils/InvertedPromise.js';
 
 // Api
 export { performSp1Plonk as performSp1ToPlonk } from './api/sp1/plonk.js';
-export { Sp1Input as Sp1 } from './api/sp1/types.js';
+export { Sp1Input as Sp1 } from './api/validation/sp1/schema.js';
 
 // Compute
 export { ComputationalPlanExecutor } from './compute/executor.js';
